# frontend/app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "paths": {
...
    "/api/v1/villages": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "in": "query",
            "name": "charachip",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "in": "query",
            "name": "skill",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "in": "query",
            "name": "random",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageListResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "dummyCharaImage": {
                    "type": "string",
                    "format": "binary"
                  },
                  "request": {
                    "$ref": "#/components/schemas/VillageCreateRequest"
                  }
                },
                "required": ["request"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageCreateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
    "/api/v1/villages/{id}": {
      "get": {
        "operationId": "getVillage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageDetailView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
...
    "/api/v1/villages/{id}/info": {
      "get": {
        "operationId": "getVillageInfo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageSettingsContent"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
...
    "/api/v1/villages/{id}/setting": {
      "get": {
        "operationId": "setting",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageSettingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
    "/api/v1/villages/{id}/situation": {
      "get": {
        "operationId": "getVillageSituation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageSituationView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
    "/api/v1/villages/{id}/situation/me": {
      "get": {
        "operationId": "getMyVillageSituation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantSituationView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
...
    "/api/v1/villages/{id}/update": {
      "post": {
        "operationId": "updateVillage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VillageUpdateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["village-rest-controller"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageTypeSayRestrict": {
        "type": "object",
        "properties": {
          "count": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "length": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "messageTypeCode": {
            "type": ["string", "null"]
          },
          "restrict": {
            "type": ["boolean", "null"]
          }
        },
        "required": ["messageTypeCode", "restrict"]
      },
...
      "VillageCreateRequest": {
        "type": "object",
        "properties": {
          "ageLimit": {
            "type": ["string", "null"],
            "pattern": "R15|R18"
          },
          "allowedSecretSayCode": {
            "type": ["string", "null"],
            "pattern": "NOTHING|ONLY_CREATOR|EVERYTHING"
          },
          "availableAction": {
            "type": ["boolean", "null"]
          },
          "availableCommit": {
            "type": ["boolean", "null"]
          },
          "availableGuardSameTarget": {
            "type": ["boolean", "null"]
          },
          "availableSameWolfAttack": {
            "type": ["boolean", "null"]
          },
          "availableSpectate": {
            "type": ["boolean", "null"]
          },
          "availableSuddonlyDeath": {
            "type": ["boolean", "null"]
          },
          "campAllocationList": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/VillageCreateRequestCampAllocation"
            }
          },
          "characterSetId": {
            "type": ["array", "null"],
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "creatorIsProducer": {
            "type": ["boolean", "null"]
          },
          "dayChangeIntervalHours": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 72,
            "minimum": 0
          },
          "dayChangeIntervalMinutes": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 59,
            "minimum": 0
          },
          "dayChangeIntervalSeconds": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 59,
            "minimum": 0
          },
          "dummyCharaId": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "dummyCharaName": {
            "type": ["string", "null"],
            "maxLength": 40,
            "minLength": 1
          },
          "dummyCharaShortName": {
            "type": ["string", "null"],
            "maxLength": 1,
            "minLength": 1
          },
          "dummyDay1Message": {
            "type": ["string", "null"],
            "maxLength": 400,
            "minLength": 0
          },
          "dummyJoinMessage": {
            "type": ["string", "null"],
            "maxLength": 400,
            "minLength": 1
          },
          "joinPassword": {
            "type": ["string", "null"]
          },
          "openSkillInGrave": {
            "type": ["boolean", "null"]
          },
          "openVote": {
            "type": ["boolean", "null"]
          },
          "organization": {
            "type": ["string", "null"]
          },
          "personMaxNum": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 999
          },
          "possibleSkillRequest": {
            "type": ["boolean", "null"]
          },
          "randomOrganization": {
            "type": ["boolean", "null"]
          },
          "reincarnationSkillAll": {
            "type": ["boolean", "null"]
          },
          "rpSayRestrictList": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/MessageTypeSayRestrict"
            }
          },
          "sayRestrictList": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/SkillSayRestrict"
            }
          },
          "shouldOriginalImage": {
            "type": ["boolean", "null"]
          },
          "skillSayRestrictList": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/MessageTypeSayRestrict"
            }
          },
          "startDay": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 31,
            "minimum": 1
          },
          "startHour": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 23,
            "minimum": 0
          },
          "startMinute": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 59,
            "minimum": 0
          },
          "startMonth": {
            "type": ["integer", "null"],
            "format": "int32",
            "maximum": 12,
            "minimum": 1
          },
          "startPersonMinNum": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 8
          },
          "startYear": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "villageName": {
            "type": ["string", "null"],
            "maxLength": 40,
            "minLength": 5
          },
          "visibleGraveSpectateMessage": {
            "type": ["boolean", "null"]
          },
          "welcomeRange": {
            "type": ["string", "null"],
            "pattern": "ANYONE_WELCOME|RELATIVES_ONLY"
          },
          "wolfAllocation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/VillageCreateRequestWolfAllocation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "allowedSecretSayCode",
          "availableAction",
          "availableCommit",
          "availableGuardSameTarget",
          "availableSameWolfAttack",
          "availableSpectate",
          "availableSuddonlyDeath",
          "characterSetId",
          "creatorIsProducer",
          "dayChangeIntervalHours",
          "dayChangeIntervalMinutes",
          "dayChangeIntervalSeconds",
          "dummyCharaName",
          "dummyCharaShortName",
          "dummyJoinMessage",
          "openSkillInGrave",
          "openVote",
          "personMaxNum",
          "possibleSkillRequest",
          "randomOrganization",
          "reincarnationSkillAll",
          "rpSayRestrictList",
          "sayRestrictList",
          "shouldOriginalImage",
          "skillSayRestrictList",
          "startDay",
          "startHour",
          "startMinute",
          "startMonth",
          "startPersonMinNum",
          "startYear",
          "villageName",
          "visibleGraveSpectateMessage"
        ]
      },
...
      "VillageCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["id"]
      },
...
      "VillageDetailView": {
        "type": "object",
        "properties": {
          "days": {
            "$ref": "#/components/schemas/VillageDays"
          },
          "epilogueDay": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "participants": {
            "$ref": "#/components/schemas/VillageParticipantsView"
          },
          "roomSize": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RoomSize"
              },
              {
                "type": "null"
              }
            ]
          },
          "setting": {
            "$ref": "#/components/schemas/VillageSettingView"
          },
          "spectators": {
            "$ref": "#/components/schemas/VillageParticipantsView"
          },
          "status": {
            "$ref": "#/components/schemas/VillageStatus"
          }
        },
        "required": ["days", "id", "name", "participants", "setting", "spectators", "status"]
      },
...
      "VillageListResponse": {
        "type": "object",
        "properties": {
          "villages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimpleVillageView"
            }
          }
        },
        "required": ["villages"]
      },
...
      "VillageSettingView": {
        "type": "object",
        "properties": {
          "chara": {
            "$ref": "#/components/schemas/VillageCharaSetting"
          },
          "dayChangeIntervalSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "hasJoinPassword": {
            "type": "boolean"
          },
          "organize": {
            "$ref": "#/components/schemas/VillageOrganize"
          },
          "personMax": {
            "type": "integer",
            "format": "int32"
          },
          "personMin": {
            "type": "integer",
            "format": "int32"
          },
          "rule": {
            "$ref": "#/components/schemas/VillageRule"
          },
          "sayRestriction": {
            "$ref": "#/components/schemas/SayRestriction"
          },
          "startDatetime": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "$ref": "#/components/schemas/VillageTags"
          }
        },
        "required": [
          "chara",
          "dayChangeIntervalSeconds",
          "hasJoinPassword",
          "organize",
          "personMax",
          "personMin",
          "rule",
          "sayRestriction",
          "startDatetime",
          "tags"
        ]
      },
...
      "VillageSettingsContent": {
        "type": "object",
        "properties": {
          "ageLimit": {
            "type": "string"
          },
          "allowedSecretSayCode": {
            "type": "string"
          },
          "campAllocationList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSettingsRandomCampOrganization"
            }
          },
          "charachips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSettingsCharachip"
            }
          },
          "createPlayerName": {
            "type": "string"
          },
          "creatorIsProducer": {
            "type": "boolean"
          },
          "dayChangeInterval": {
            "type": "string"
          },
          "dummyCharaName": {
            "type": "string"
          },
          "isAvailableAction": {
            "type": "boolean"
          },
          "isAvailableCommit": {
            "type": "boolean"
          },
          "isAvailableGuardSameTarget": {
            "type": "boolean"
          },
          "isAvailableSameWolfAttack": {
            "type": "boolean"
          },
          "isAvailableSpectate": {
            "type": "boolean"
          },
          "isAvailableSuddenlyDeath": {
            "type": "boolean"
          },
          "isOpenSkillInGrave": {
            "type": "boolean"
          },
          "isRandomOrganization": {
            "type": "boolean"
          },
          "isReincarnationSkillAll": {
            "type": "boolean"
          },
          "isRequiredJoinPassword": {
            "type": "boolean"
          },
          "isSkillRequestAvailable": {
            "type": "boolean"
          },
          "isVisibleGraveSpectateMessage": {
            "type": "boolean"
          },
          "organization": {
            "type": "string"
          },
          "personMaxNum": {
            "type": "integer",
            "format": "int32"
          },
          "rpSayRestrictList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSettingsRpSayRestriction"
            }
          },
          "sayRestrictList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSettingsSayRestriction"
            }
          },
          "shouldOriginalImage": {
            "type": "boolean"
          },
          "skillRequestType": {
            "type": "string"
          },
          "skillSayRestrictList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSettingsSkillSayRestriction"
            }
          },
          "startDatetime": {
            "type": "string",
            "format": "date-time"
          },
          "startPersonMinNum": {
            "type": "integer",
            "format": "int32"
          },
          "voteType": {
            "type": "string"
          },
          "welcomeRange": {
            "type": ["string", "null"]
          },
          "wolfAllocation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/VillageSettingsRandomWolfOrganization"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ageLimit",
          "allowedSecretSayCode",
          "campAllocationList",
          "charachips",
          "createPlayerName",
          "creatorIsProducer",
          "dayChangeInterval",
          "dummyCharaName",
          "isAvailableAction",
          "isAvailableCommit",
          "isAvailableGuardSameTarget",
          "isAvailableSameWolfAttack",
          "isAvailableSpectate",
          "isAvailableSuddenlyDeath",
          "isOpenSkillInGrave",
          "isRandomOrganization",
          "isReincarnationSkillAll",
          "isRequiredJoinPassword",
          "isSkillRequestAvailable",
          "isVisibleGraveSpectateMessage",
          "organization",
          "personMaxNum",
          "rpSayRestrictList",
          "sayRestrictList",
          "shouldOriginalImage",
          "skillRequestType",
          "skillSayRestrictList",
          "startDatetime",
          "startPersonMinNum",
          "voteType"
        ]
      },
...
      "VillageSituationView": {
        "type": "object",
        "properties": {
          "footstepList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageFootstepContent"
            }
          },
          "isViewableSpoilerContent": {
            "type": "boolean"
          },
          "memberList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageMemberContent"
            }
          },
          "participantList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageFilterParticipantContent"
            }
          },
          "roomAssignedRowList": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/VillageRoomAssignedRow"
            }
          },
          "roomWidth": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "situationList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillageSituationContent"
            }
          },
          "vote": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/VillageVoteContent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "footstepList",
          "isViewableSpoilerContent",
          "memberList",
          "participantList",
          "situationList"
        ]
      },
...
      "VillageUpdateResponse": {
        "type": "object",
        "properties": {
          "latestDay": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["latestDay"]
      },
...
    }
  }
}